requirements.txt
streamlit==1.30.0
numpy==1.26.4
Pillow==10.2.0
tensorflow==2.16.1
scikit-learn==1.4.1.post1
matplotlib==3.8.3
seaborn==0.13.2